an untested MySensors library too big to fit in 8kB

+++ hardware/stm8/0.5.0/libraries/EEPROM/src/EEPROM_update.c
@@ -29,7 +29,7 @@
 
 /* --- Arduino-like interface -------------------------------------------- */
 
-void EEPROM_update( int idx, uint8_t val )
+void EEPROM_update( uint16_t idx, uint8_t val )
 {
 	// make sure not to write data beyond the end of the EEPROM area
 	// (this could accidentally hit the option byte area)
